# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /weekly-orders:
    get:
      tags: [WeeklyOrders]
      operationId: listWeeklyOrders
      summary: List every school's order for a given week
      parameters:
        - name: weekStart
          in: query
          required: true
          description: Monday date of the target week (YYYY-MM-DD).
          schema:
            type: string
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '200':
          description: One row per school for the chosen week.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WeeklyOrderSummary'
  /weekly-orders/{id}:
    get:
      tags: [WeeklyOrders]
      operationId: getWeeklyOrder
      summary: Get an existing weekly order with items
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '200':
          description: Weekly order.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WeeklyOrder'
    patch:
      tags: [WeeklyOrders]
      operationId: updateWeeklyOrder
      summary: Update a weekly order's status or notes (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateWeeklyOrderBody'
      responses:
        '200':
          description: Updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WeeklyOrder'
  /weekly-orders/open-or-create:
    post:
      tags: [WeeklyOrders]
      operationId: openOrCreateWeeklyOrder
      summary: Open or create the weekly order row for a school for a given week
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OpenWeeklyOrderBody'
      responses:
        '200':
          description: Weekly order, possibly newly created from defaults.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WeeklyOrder'
  /weekly-orders/{id}/items:
    post:
      tags: [WeeklyOrders]
      operationId: addWeeklyOrderItem
      summary: Add or update an item on a weekly order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertWeeklyOrderItemBody'
      responses:
        '200':
          description: Item upserted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WeeklyOrderItem'
  /weekly-orders/{id}/items/{itemId}:
    patch:
      tags: [WeeklyOrders]
      operationId: updateWeeklyOrderItem
      summary: Update a weekly order item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: itemId
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateWeeklyOrderItemBody'
      responses:
        '200':
          description: Updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WeeklyOrderItem'
    delete:
      tags: [WeeklyOrders]
      operationId: removeWeeklyOrderItem
      summary: Remove an item from a weekly order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: itemId
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '204':
          description: Removed.
...
components:
  parameters:
    AuthorizationSessionHeader:
      name: Authorization
      in: header
      required: false
      description: Opaque session token — `Bearer <sid>`.
      schema:
        type: string
    SessionCookie:
      name: sid
      in: cookie
      required: false
      description: Browser session cookie.
      schema:
        type: string
  schemas:
...
    Product:
      type: object
      required:
        - id
        - name
        - description
        - category
        - unit
        - sku
        - allergens
        - active
        - minInventory
        - maxInventory
        - createdAt
        - updatedAt
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: [string, 'null'] }
        category: { type: [string, 'null'] }
        unit: { type: [string, 'null'] }
        sku: { type: [string, 'null'] }
        allergens: { type: [string, 'null'] }
        active: { type: boolean }
        minInventory: { type: [integer, 'null'], minimum: 0 }
        maxInventory: { type: [integer, 'null'], minimum: 0 }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    WeeklyOrderItem:
      type: object
      required: [id, orderId, productId, quantity, note, product]
      properties:
        id: { type: string }
        orderId: { type: string }
        productId: { type: string }
        quantity: { type: integer, minimum: 0 }
        note: { type: [string, 'null'] }
        product:
          $ref: '#/components/schemas/Product'
    WeeklyOrder:
      type: object
      required:
        - id
        - schoolId
        - schoolName
        - weekStart
        - status
        - notes
        - confirmedAt
        - items
        - createdAt
        - updatedAt
      properties:
        id: { type: string }
        schoolId: { type: string }
        schoolName: { type: string }
        weekStart: { type: string }
        status:
          type: string
          enum: [not_started, in_progress, confirmed]
        notes: { type: [string, 'null'] }
        confirmedAt: { type: [string, 'null'], format: date-time }
        items:
          type: array
          items:
            $ref: '#/components/schemas/WeeklyOrderItem'
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    WeeklyOrderSummary:
      type: object
      required:
        - schoolId
        - schoolName
        - weekStart
        - status
        - itemCount
        - notesPreview
      properties:
        orderId: { type: [string, 'null'] }
        schoolId: { type: string }
        schoolName: { type: string }
        weekStart: { type: string }
        status:
          type: string
          enum: [not_started, in_progress, confirmed]
        itemCount: { type: integer, minimum: 0 }
        notesPreview: { type: [string, 'null'] }
        confirmedAt: { type: [string, 'null'], format: date-time }
        routeWeekInstanceId: { type: [string, 'null'] }
        routeName: { type: [string, 'null'] }
        truckName: { type: [string, 'null'] }
        dayOfWeek: { type: [integer, 'null'] }
        driverName: { type: [string, 'null'] }
    UpdateWeeklyOrderBody:
      type: object
      properties:
        notes: { type: [string, 'null'] }
        status:
          type: string
          enum: [not_started, in_progress, confirmed]
    OpenWeeklyOrderBody:
      type: object
      required: [schoolId, weekStart]
      properties:
        schoolId: { type: string, minLength: 1 }
        weekStart: { type: string, minLength: 1 }
    UpsertWeeklyOrderItemBody:
      type: object
      required: [productId, quantity]
      properties:
        productId: { type: string, minLength: 1 }
        quantity: { type: integer, minimum: 0 }
        note: { type: [string, 'null'] }
    UpdateWeeklyOrderItemBody:
      type: object
      properties:
        quantity: { type: integer, minimum: 0 }
        note: { type: [string, 'null'] }
...