# lib/api-zod/src/generated/types/weeklyOrderSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { WeeklyOrderSummaryStatus } from "./weeklyOrderSummaryStatus";

export interface WeeklyOrderSummary {
  /** @nullable */
  orderId?: string | null;
  schoolId: string;
  schoolName: string;
  weekStart: string;
  status: WeeklyOrderSummaryStatus;
  /** @minimum 0 */
  itemCount: number;
  /** @nullable */
  notesPreview: string | null;
  /** @nullable */
  confirmedAt?: Date | null;
  /** @nullable */
  routeWeekInstanceId?: string | null;
  /** @nullable */
  routeName?: string | null;
  /** @nullable */
  truckName?: string | null;
  /** @nullable */
  dayOfWeek?: number | null;
  /** @nullable */
  driverName?: string | null;
}